requirements.txt
azure-search-documents==11.3.0b4
requests
